change method and data type

+++ src/main/java/co/id/Asset/eInvoice/Database/Repository/DescriptionRepository.java
@@ -4,6 +4,6 @@ import co.id.Asset.eInvoice.Database.Entity.Description;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface DescriptionRepository extends JpaRepository<Description, Long> {
+public interface DescriptionRepository extends JpaRepository<Description, Integer> {
     List<Description> findByInvoiceNumber(String invoiceNumber);
 }
